Add files via upload
I don't think this is the particular format that we have to turn in Milestone 4 but this file contains the code I've worked on - writing the queries for making the new Symptoms table and adjusting all the Join queries from Milestone 3 to work for Symptoms and Vaccines tables in dataset1clean.  I'm almost sure that all of those queries will work when we have Vaccines table in there with all the fields that we talked about.